refactor 2 of App.tsx into seperate components
